Melhorias na controller e model do produto

+++ controller/productController.js
@@ -1,45 +1,45 @@
 const fs = require("fs");
-const Product = require('../models/Product')
+const Product = require('../models/Product');
 const upload = require('../config/multer');
 
 function productController(app) {
     app.post("/product/create", upload.array("images", 5), async (req, res) => {
-        if (!req.files || req.files.length !== 5) {
+        if (!req.files || (req.files.length !== 4 && req.files.length !== 5)) {
             if (req.files) {
                 req.files.forEach(file => fs.unlink(file.path, err => {
                     if (err) console.error(`Erro ao deletar o arquivo: ${file.path}`);
                 }));
             }
-            return res.status(400).json({ msg: "Exatamente 5 imagens são necessárias" });
+            return res.status(400).json({ msg: "São necessárias 4 ou 5 imagens" });
         }
 
         const {
             nome,
             precoPrazo,
-            preco,
             descricao,
             especificacoes,
             marca,
             categoria,
-            estoque
         } = req.body;
 
         const imgPaths = req.files.map(file => file.path);
 
+        const images = imgPaths; 
+
+        if (req.files.length === 5) {
+            images.img5 = imgPaths[4];
+        }
+
         const product = new Product({
             nome,
             precoPrazo,
-            preco,
+            preco: precoPrazo - precoPrazo * 0.1,
             descricao,
             especificacoes,
             marca,
             categoria,
-            img1: imgPaths[0],
-            img2: imgPaths[1],
-            img3: imgPaths[2],
-            img4: imgPaths[3],
-            img5: imgPaths[4],
-            estoque
+            images,
+            estoque: 0
         });
 
         try {
@@ -65,6 +65,21 @@ function productController(app) {
             res.status(500).json({ msg: "Erro no servidor!" });
         }
     });
+
+    app.get("/product/", async (req, res) => {
+
+        try {
+            const products = await Product.find();
+            if (products.length === 0) {
+                return res.status(404).json({ msg: "Nenhum produto encontrado!" });
+            }
+            res.status(200).json({ products });
+        } catch (error) {
+            console.error(error);
+            res.status(500).json({ msg: "Erro no servidor!" });
+        }
+    });
+
 }
 
 module.exports = { productController };
